requirements.txt
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
cryptography==45.0.4
distlib==0.3.9
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.1.0
fastapi==0.104.1
filelock==3.18.0
Flask==3.1.1
greenlet==3.2.3
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
passlib==1.7.4
platformdirs==4.3.8
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
python-jose==3.3.0
python-multipart==0.0.6
requests==2.32.4
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.27.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.24.0
virtualenv==20.31.2
Werkzeug==3.1.3
